+++ modules/evm-bridge/src/mock.rs
@@ -109,7 +109,7 @@ impl module_evm::Config for Runtime {
 	type PrecompilesValue = ();
 	type ChainId = ();
 	type GasToWeight = ();
-	type ChargeTransactionPayment = support::MockReservedTransactionPayment<Balances>;
+	type ChargeTransactionPayment = support::mocks::MockReservedTransactionPayment<Balances>;
 	type NetworkContractOrigin = EnsureSignedBy<NetworkContractAccount, AccountId32>;
 	type NetworkContractSource = NetworkContractSource;
