+++ modules/support/src/lib.rs
@@ -102,102 +102,6 @@ pub trait TransactionPayment<AccountId, Balance, NegativeImbalance> {
 	) -> Result<(), TransactionValidityError>;
 }
 
-#[cfg(feature = "std")]
-use frame_support::traits::Imbalance;
-#[cfg(feature = "std")]
-impl<AccountId, Balance: Default + Copy, NegativeImbalance: Imbalance<Balance>>
-	TransactionPayment<AccountId, Balance, NegativeImbalance> for ()
-{
-	fn reserve_fee(
-		_who: &AccountId,
-		_fee: Balance,
-		_named: Option<ReserveIdentifier>,
-	) -> Result<Balance, DispatchError> {
-		Ok(Default::default())
-	}
-
-	fn unreserve_fee(_who: &AccountId, _fee: Balance, _named: Option<ReserveIdentifier>) -> Balance {
-		Default::default()
-	}
-
-	fn unreserve_and_charge_fee(
-		_who: &AccountId,
-		_weight: Weight,
-	) -> Result<(Balance, NegativeImbalance), TransactionValidityError> {
-		Ok((Default::default(), Imbalance::zero()))
-	}
-
-	fn refund_fee(
-		_who: &AccountId,
-		_weight: Weight,
-		_payed: NegativeImbalance,
-	) -> Result<(), TransactionValidityError> {
-		Ok(())
-	}
-
-	fn charge_fee(
-		_who: &AccountId,
-		_len: u32,
-		_weight: Weight,
-		_tip: Balance,
-		_pays_fee: Pays,
-		_class: DispatchClass,
-	) -> Result<(), TransactionValidityError> {
-		Ok(())
-	}
-}
-
-/// Given provided `Currency`, implements default reserve behavior
-pub struct MockReservedTransactionPayment<Currency>(PhantomData<Currency>);
-
-#[cfg(feature = "std")]
-impl<
-		AccountId,
-		Balance: Default + Copy,
-		NegativeImbalance: Imbalance<Balance>,
-		Currency: frame_support::traits::NamedReservableCurrency<
-			AccountId,
-			ReserveIdentifier = ReserveIdentifier,
-			Balance = Balance,
-		>,
-	> TransactionPayment<AccountId, Balance, NegativeImbalance> for MockReservedTransactionPayment<Currency>
-{
-	fn reserve_fee(who: &AccountId, fee: Balance, named: Option<ReserveIdentifier>) -> Result<Balance, DispatchError> {
-		Currency::reserve_named(&named.unwrap(), who, fee)?;
-		Ok(fee)
-	}
-
-	fn unreserve_fee(who: &AccountId, fee: Balance, named: Option<ReserveIdentifier>) -> Balance {
-		Currency::unreserve_named(&named.unwrap(), who, fee)
-	}
-
-	fn unreserve_and_charge_fee(
-		_who: &AccountId,
-		_weight: Weight,
-	) -> Result<(Balance, NegativeImbalance), TransactionValidityError> {
-		Ok((Default::default(), Imbalance::zero()))
-	}
-
-	fn refund_fee(
-		_who: &AccountId,
-		_weight: Weight,
-		_payed: NegativeImbalance,
-	) -> Result<(), TransactionValidityError> {
-		Ok(())
-	}
-
-	fn charge_fee(
-		_who: &AccountId,
-		_len: u32,
-		_weight: Weight,
-		_tip: Balance,
-		_pays_fee: Pays,
-		_class: DispatchClass,
-	) -> Result<(), TransactionValidityError> {
-		Ok(())
-	}
-}
-
 /// Used to interface with the Compound's Cash module
 pub trait CompoundCashTrait<Balance, Moment> {
 	fn set_future_yield(next_cash_yield: Balance, yield_index: u128, timestamp_effective: Moment) -> DispatchResult;
