+++ modules/support/src/mocks.rs
@@ -17,20 +17,24 @@
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
 #![allow(clippy::type_complexity)]
-use crate::{AddressMapping, CurrencyId, Erc20InfoMapping};
+use crate::{AddressMapping, CurrencyId, Erc20InfoMapping, TransactionPayment};
 use codec::Encode;
-use nutsfinance_stable_asset::traits::StableAsset;
+use frame_support::pallet_prelude::{DispatchClass, Pays, Weight};
 use nutsfinance_stable_asset::{
-	PoolTokenIndex, RedeemProportionResult, StableAssetPoolId, StableAssetPoolInfo, SwapResult,
+	traits::StableAsset, PoolTokenIndex, RedeemProportionResult, StableAssetPoolId, StableAssetPoolInfo, SwapResult,
 };
 use primitives::{
 	currency::TokenInfo,
 	evm::{EvmAddress, H160_POSITION_TOKEN},
+	ReserveIdentifier,
 };
 use sp_core::{crypto::AccountId32, H160};
 use sp_io::hashing::blake2_256;
-use sp_runtime::{DispatchError, DispatchResult};
-use sp_std::vec::Vec;
+use sp_runtime::{transaction_validity::TransactionValidityError, DispatchError, DispatchResult};
+use sp_std::{marker::PhantomData, vec::Vec};
+
+#[cfg(feature = "std")]
+use frame_support::traits::Imbalance;
 
 pub struct MockAddressMapping;
 
@@ -95,6 +99,100 @@ impl Erc20InfoMapping for MockErc20InfoMapping {
 		EvmAddress::try_from(token)
 			.map(|addr| if addr == v { Some(token) } else { None })
 			.ok()?
+	}
+}
+
+#[cfg(feature = "std")]
+impl<AccountId, Balance: Default + Copy, NegativeImbalance: Imbalance<Balance>>
+	TransactionPayment<AccountId, Balance, NegativeImbalance> for ()
+{
+	fn reserve_fee(
+		_who: &AccountId,
+		_fee: Balance,
+		_named: Option<ReserveIdentifier>,
+	) -> Result<Balance, DispatchError> {
+		Ok(Default::default())
+	}
+
+	fn unreserve_fee(_who: &AccountId, _fee: Balance, _named: Option<ReserveIdentifier>) -> Balance {
+		Default::default()
+	}
+
+	fn unreserve_and_charge_fee(
+		_who: &AccountId,
+		_weight: Weight,
+	) -> Result<(Balance, NegativeImbalance), TransactionValidityError> {
+		Ok((Default::default(), Imbalance::zero()))
+	}
+
+	fn refund_fee(
+		_who: &AccountId,
+		_weight: Weight,
+		_payed: NegativeImbalance,
+	) -> Result<(), TransactionValidityError> {
+		Ok(())
+	}
+
+	fn charge_fee(
+		_who: &AccountId,
+		_len: u32,
+		_weight: Weight,
+		_tip: Balance,
+		_pays_fee: Pays,
+		_class: DispatchClass,
+	) -> Result<(), TransactionValidityError> {
+		Ok(())
+	}
+}
+
+/// Given provided `Currency`, implements default reserve behavior
+pub struct MockReservedTransactionPayment<Currency>(PhantomData<Currency>);
+
+#[cfg(feature = "std")]
+impl<
+		AccountId,
+		Balance: Default + Copy,
+		NegativeImbalance: Imbalance<Balance>,
+		Currency: frame_support::traits::NamedReservableCurrency<
+			AccountId,
+			ReserveIdentifier = ReserveIdentifier,
+			Balance = Balance,
+		>,
+	> TransactionPayment<AccountId, Balance, NegativeImbalance> for MockReservedTransactionPayment<Currency>
+{
+	fn reserve_fee(who: &AccountId, fee: Balance, named: Option<ReserveIdentifier>) -> Result<Balance, DispatchError> {
+		Currency::reserve_named(&named.unwrap(), who, fee)?;
+		Ok(fee)
+	}
+
+	fn unreserve_fee(who: &AccountId, fee: Balance, named: Option<ReserveIdentifier>) -> Balance {
+		Currency::unreserve_named(&named.unwrap(), who, fee)
+	}
+
+	fn unreserve_and_charge_fee(
+		_who: &AccountId,
+		_weight: Weight,
+	) -> Result<(Balance, NegativeImbalance), TransactionValidityError> {
+		Ok((Default::default(), Imbalance::zero()))
+	}
+
+	fn refund_fee(
+		_who: &AccountId,
+		_weight: Weight,
+		_payed: NegativeImbalance,
+	) -> Result<(), TransactionValidityError> {
+		Ok(())
+	}
+
+	fn charge_fee(
+		_who: &AccountId,
+		_len: u32,
+		_weight: Weight,
+		_tip: Balance,
+		_pays_fee: Pays,
+		_class: DispatchClass,
+	) -> Result<(), TransactionValidityError> {
+		Ok(())
 	}
 }
